# docs/contracts/openapi/memory_agent-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Memory Agent",
    "description": "\u0648\u0643\u064a\u0644 \u0645\u0633\u062a\u0642\u0644 \u0644\u0625\u062f\u0627\u0631\u0629 \u0627\u0644\u0633\u064a\u0627\u0642 \u0648\u0627\u0644\u0630\u0627\u0643\u0631\u0629",
    "version": "1.0.0"
  },
  "paths": {
...
    "/knowledge/concepts/search": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Find Concept",
        "description": "\u064a\u0628\u062d\u062b \u0639\u0646 \u0645\u0641\u0647\u0648\u0645.",
        "operationId": "find_concept_knowledge_concepts_search_get",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "\u0627\u0644\u0645\u0648\u0636\u0648\u0639 \u0644\u0644\u0628\u062d\u062b \u0639\u0646\u0647",
              "title": "Topic"
            },
            "description": "\u0627\u0644\u0645\u0648\u0636\u0648\u0639 \u0644\u0644\u0628\u062d\u062b \u0639\u0646\u0647"
          },
          {
            "name": "X-Service-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Concept"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/concepts/{concept_id}": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get Concept",
        "description": "\u064a\u062d\u0635\u0644 \u0639\u0644\u0649 \u0645\u0641\u0647\u0648\u0645 \u0628\u0648\u0627\u0633\u0637\u0629 \u0645\u0639\u0631\u0641\u0647.",
        "operationId": "get_concept_knowledge_concepts__concept_id__get",
        "parameters": [
          {
            "name": "concept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Concept Id"
            }
          },
          {
            "name": "X-Service-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Concept"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/concepts/{concept_id}/prerequisites": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get Prerequisites",
        "description": "\u064a\u062d\u0635\u0644 \u0639\u0644\u0649 \u0627\u0644\u0645\u062a\u0637\u0644\u0628\u0627\u062a \u0627\u0644\u0633\u0627\u0628\u0642\u0629.",
        "operationId": "get_prerequisites_knowledge_concepts__concept_id__prerequisites_get",
        "parameters": [
          {
            "name": "concept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Concept Id"
            }
          },
          {
            "name": "X-Service-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Concept"
                  },
                  "title": "Response Get Prerequisites Knowledge Concepts  Concept Id  Prerequisites Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/concepts/{concept_id}/related": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get Related",
        "description": "\u064a\u062d\u0635\u0644 \u0639\u0644\u0649 \u0627\u0644\u0645\u0641\u0627\u0647\u064a\u0645 \u0627\u0644\u0645\u0631\u062a\u0628\u0637\u0629.",
        "operationId": "get_related_knowledge_concepts__concept_id__related_get",
        "parameters": [
          {
            "name": "concept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Concept Id"
            }
          },
          {
            "name": "X-Service-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Concept"
                  },
                  "title": "Response Get Related Knowledge Concepts  Concept Id  Related Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/concepts/{concept_id}/next": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get Next",
        "description": "\u064a\u062d\u0635\u0644 \u0639\u0644\u0649 \u0627\u0644\u0645\u0641\u0627\u0647\u064a\u0645 \u0627\u0644\u062a\u0627\u0644\u064a\u0629.",
        "operationId": "get_next_knowledge_concepts__concept_id__next_get",
        "parameters": [
          {
            "name": "concept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Concept Id"
            }
          },
          {
            "name": "X-Service-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Concept"
                  },
                  "title": "Response Get Next Knowledge Concepts  Concept Id  Next Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/paths": {
      "post": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get Learning Path",
        "description": "\u064a\u062c\u062f \u0645\u0633\u0627\u0631 \u062a\u0639\u0644\u0645.",
        "operationId": "get_learning_path_knowledge_paths_post",
        "parameters": [
          {
            "name": "X-Service-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Concept"
                  },
                  "title": "Response Get Learning Path Knowledge Paths Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/concepts/batch/relations": {
      "post": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get Batch Relations",
        "description": "\u064a\u0633\u062a\u062e\u0631\u062c \u0627\u0644\u0639\u0644\u0627\u0642\u0627\u062a \u0644\u0645\u062c\u0645\u0648\u0639\u0629 \u0645\u0646 \u0627\u0644\u0645\u0641\u0627\u0647\u064a\u0645 \u062f\u0641\u0639\u0629 \u0648\u0627\u062d\u062f\u0629.",
        "operationId": "get_batch_relations_knowledge_concepts_batch_relations_post",
        "parameters": [
          {
            "name": "X-Service-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRelationsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRelationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/readiness": {
      "post": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Check Readiness",
        "description": "\u064a\u062a\u062d\u0642\u0642 \u0645\u0646 \u062c\u0627\u0647\u0632\u064a\u0629 \u0627\u0644\u0637\u0627\u0644\u0628 \u0644\u0645\u0641\u0647\u0648\u0645 \u0645\u0639\u064a\u0646.",
        "operationId": "check_readiness_knowledge_readiness_post",
        "parameters": [
          {
            "name": "X-Service-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadinessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BatchRelationsRequest": {
        "properties": {
          "concept_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Concept Ids",
            "description": "\u0642\u0627\u0626\u0645\u0629 \u0645\u0639\u0631\u0641\u0627\u062a \u0627\u0644\u0645\u0641\u0627\u0647\u064a\u0645"
          }
        },
        "type": "object",
        "required": [
          "concept_ids"
        ],
        "title": "BatchRelationsRequest",
        "description": "\u0637\u0644\u0628 \u0641\u062d\u0635 \u0627\u0644\u0639\u0644\u0627\u0642\u0627\u062a \u0628\u0627\u0644\u062c\u0645\u0644\u0629."
      },
      "BatchRelationsResponse": {
        "properties": {
          "prerequisites": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Prerequisites",
            "description": "\u0627\u0644\u0645\u062a\u0637\u0644\u0628\u0627\u062a \u0627\u0644\u0633\u0627\u0628\u0642\u0629 \u0644\u0643\u0644 \u0645\u0641\u0647\u0648\u0645 (\u0627\u0644\u0645\u0641\u0627\u062a\u064a\u062d \u0647\u064a concept_id\u060c \u0648\u0627\u0644\u0642\u064a\u0645 \u0647\u064a \u0642\u0627\u0626\u0645\u0629 \u0627\u0644\u0645\u062a\u0637\u0644\u0628\u0627\u062a)"
          }
        },
        "type": "object",
        "required": [
          "prerequisites"
        ],
        "title": "BatchRelationsResponse",
        "description": "\u0627\u0633\u062a\u062c\u0627\u0628\u0629 \u0627\u0644\u0639\u0644\u0627\u0642\u0627\u062a \u0628\u0627\u0644\u062c\u0645\u0644\u0629."
      },
      "Concept": {
        "properties": {
          "concept_id": {
            "type": "string",
            "title": "Concept Id",
            "description": "\u0645\u0639\u0631\u0641 \u0627\u0644\u0645\u0641\u0647\u0648\u0645"
          },
          "name_ar": {
            "type": "string",
            "title": "Name Ar",
            "description": "\u0627\u0644\u0627\u0633\u0645 \u0628\u0627\u0644\u0639\u0631\u0628\u064a\u0629"
          },
          "name_en": {
            "type": "string",
            "title": "Name En",
            "description": "\u0627\u0644\u0627\u0633\u0645 \u0628\u0627\u0644\u0625\u0646\u062c\u0644\u064a\u0632\u064a\u0629",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "\u0648\u0635\u0641 \u0627\u0644\u0645\u0641\u0647\u0648\u0645",
            "default": ""
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "description": "\u0627\u0644\u0645\u0627\u062f\u0629",
            "default": "Mathematics"
          },
          "level": {
            "type": "string",
            "title": "Level",
            "description": "\u0627\u0644\u0645\u0633\u062a\u0648\u0649",
            "default": ""
          },
          "difficulty": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Difficulty",
            "description": "\u0627\u0644\u0635\u0639\u0648\u0628\u0629",
            "default": 0.5
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "name_ar"
        ],
        "title": "Concept",
        "description": "\u0645\u0641\u0647\u0648\u0645 \u062a\u0639\u0644\u064a\u0645\u064a."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PathRequest": {
        "properties": {
          "from_concept": {
            "type": "string",
            "title": "From Concept"
          },
          "to_concept": {
            "type": "string",
            "title": "To Concept"
          }
        },
        "type": "object",
        "required": [
          "from_concept",
          "to_concept"
        ],
        "title": "PathRequest"
      },
      "ReadinessRequest": {
        "properties": {
          "concept_id": {
            "type": "string",
            "title": "Concept Id",
            "description": "\u0645\u0639\u0631\u0641 \u0627\u0644\u0645\u0641\u0647\u0648\u0645 \u0627\u0644\u0645\u0631\u0627\u062f \u062a\u0639\u0644\u0645\u0647"
          },
          "mastery_levels": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Mastery Levels",
            "description": "\u062e\u0631\u064a\u0637\u0629 \u0645\u0633\u062a\u0648\u0649 \u0625\u062a\u0642\u0627\u0646 \u0627\u0644\u0645\u0641\u0627\u0647\u064a\u0645 (concept_id -> score)"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "mastery_levels"
        ],
        "title": "ReadinessRequest",
        "description": "\u0637\u0644\u0628 \u0641\u062d\u0635 \u0627\u0644\u062c\u0627\u0647\u0632\u064a\u0629."
      },
      "ReadinessResponse": {
        "properties": {
          "concept_id": {
            "type": "string",
            "title": "Concept Id",
            "description": "\u0645\u0639\u0631\u0641 \u0627\u0644\u0645\u0641\u0647\u0648\u0645"
          },
          "concept_name": {
            "type": "string",
            "title": "Concept Name",
            "description": "\u0627\u0633\u0645 \u0627\u0644\u0645\u0641\u0647\u0648\u0645"
          },
          "is_ready": {
            "type": "boolean",
            "title": "Is Ready",
            "description": "\u0647\u0644 \u0627\u0644\u0637\u0627\u0644\u0628 \u062c\u0627\u0647\u0632\u061f"
          },
          "readiness_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Readiness Score",
            "description": "\u062f\u0631\u062c\u0629 \u0627\u0644\u062c\u0627\u0647\u0632\u064a\u0629"
          },
          "missing_prerequisites": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Prerequisites",
            "description": "\u0623\u0633\u0645\u0627\u0621 \u0627\u0644\u0645\u062a\u0637\u0644\u0628\u0627\u062a \u0627\u0644\u0645\u0641\u0642\u0648\u062f\u0629 \u062a\u0645\u0627\u0645\u0627\u064b"
          },
          "weak_prerequisites": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Prerequisites",
            "description": "\u0623\u0633\u0645\u0627\u0621 \u0627\u0644\u0645\u062a\u0637\u0644\u0628\u0627\u062a \u0627\u0644\u0636\u0639\u064a\u0641\u0629"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation",
            "description": "\u062a\u0648\u0635\u064a\u0629 \u0644\u0644\u0637\u0627\u0644\u0628"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "concept_name",
          "is_ready",
          "readiness_score",
          "missing_prerequisites",
          "weak_prerequisites",
          "recommendation"
        ],
        "title": "ReadinessResponse",
        "description": "\u0646\u062a\u064a\u062c\u0629 \u0641\u062d\u0635 \u0627\u0644\u062c\u0627\u0647\u0632\u064a\u0629."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}